Add GetAvailableVehicles(tested)

+++ backend/CarRentalService/CarRentalService.Infrastructure/EF/Models/VehicleReadModel.cs
@@ -4,7 +4,7 @@
     {
         public string Brand { get; set; }
         public string Model { get; set; }
-        public double PricePerDay { get; set; }
+        public decimal PricePerDay { get; set; }
         public string Type { get; set; }
         public string LicensePlate { get; set; }
         public int Year { get; set; }
